+++ packages/ai-runtime/src/types/common/index.ts
@@ -29,3 +29,20 @@ export type ServiceModePermissionMap = {
   ensurePromptsEditorRoles: string[] | undefined;
   ensurePromptsAdminRoles: string[] | undefined;
 };
+
+export interface ImageModelInfo extends ModelInfoBase {
+  brand: string;
+  model: string;
+  nMin?: number;
+  nMax?: number;
+  nDefault?: number;
+  disabled?: boolean;
+  quality?: string[];
+  qualityDefault?: 'standard' | 'hd';
+  responseFormat?: string[];
+  responseFormatDefault?: 'url' | 'b64_json';
+  size?: string[];
+  sizeDefault?: '256x256' | '512x512' | '1024x1024' | '1024x1792' | '1792x1024';
+  style?: string[];
+  styleDefault?: 'vivid' | 'natural';
+}
